+++ Test/PublishPowerShellScript.Tests.ps1
@@ -325,7 +325,6 @@ function WhenPublishing
         }
     }
 
-    Import-WhiskeyTestModule -Name 'PackageManagement','PowerShellGet'
     Mock -CommandName 'Get-PackageProvider' -ModuleName 'Whiskey'
 
     $Global:Error.Clear()