+++ Whiskey/Tasks/Version.ps1
@@ -3,7 +3,7 @@
 {
     [CmdletBinding()]
     [Whiskey.Task('Version')]
-    [Whiskey.RequiresPowerShellModule('PackageManagement', Version='1.4.7',
+    [Whiskey.RequiresPowerShellModule('PackageManagement', Version='1.*',
         VersionParameterName='PackageManagementVersion')]
     param(
         [Parameter(Mandatory)]